Make sure to pick up LDFLAGS from environment.
Adapted the Makefiles for the client and server to make sure that
any LDFLAGS settings that exist in the environment are also picked up.
This brings it in line with the CXXFLAGS implementation which also
picks up settings from the environment.
This is needed because some distributions allow to set general compiler
and linker flags in a central configuration file.